% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{barycentric}
\alias{barycentric}
\title{Compute Barycentric Coordinates}
\usage{
barycentric(p, v)
}
\arguments{
\item{p}{a list containing equal length numeric vectors containing the
coordinates of the points that we want to compute barycentric coordinates
for.}

\item{v}{numeric list-matrix where each row corresponds to a triangle vertex
data and each column the x and y coordinates for each vertex.  Each element
of the underlying vectors is matched position wise to the corresponding
point in \code{p}, so all vectors in \code{p} and \code{v} are expected to be the same
length.}
}
\value{
a list with there numeric vectors containing the barycentric
coordinates.
}
\description{
Compute Barycentric Coordinates
}
