% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{beersheva_build}
\alias{beersheva_build}
\title{Polygonal layer of 376 buildings in Beer-Sheva}
\format{A \code{SpatialPolygonsDataFrame} with 10 features and 4 attributes:
\describe{
  \item{build_id}{Building ID}
  \item{floors}{Number of floors for building}
  \item{apartments}{Number of apartments}
  \item{height_m}{Building height, in meters}
  \item{elev}{Elevation above sea level of building base, in meters}
}}
\usage{
beersheva_build
}
\description{
A \code{SpatialPolygonsDataFrame} object representing the outlines of 367 buildings in the Ramot neighborhood, Beer-Sheva. The attribute \code{height_m} contains building height, in meters.
}
\keyword{datasets}
