% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boston_block}
\alias{boston_block}
\title{Polygonal layer of a building block in Boston}
\format{
A \code{SpatialPolygons} with a single feature.
}
\usage{
boston_block
}
\description{
A \code{SpatialPolygons} object representing the boundaries of a building block in Central Boston.
}
\keyword{datasets}
