% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boston_build}
\alias{boston_build}
\title{Polygonal layer of three buildings in Boston}
\format{A \code{SpatialPolygonsDataFrame} with 10 features and 4 attributes:
\describe{
  \item{objectid}{Building part ID}
  \item{build_id}{Building ID}
  \item{part_floor}{Number of floors for part}
  \item{height_m}{Building height, in meters}
}}
\usage{
boston_build
}
\description{
A \code{SpatialPolygonsDataFrame} object representing the outlines of three buildings located in Central Boston. The attribute \code{height_m} contains building height, in meters.
}
\keyword{datasets}
