% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boston_park}
\alias{boston_park}
\title{Polygonal layer of a park in Boston}
\format{A \code{SpatialPolygons} with a single feature.}
\usage{
boston_park
}
\description{
A \code{SpatialPolygons} object representing the boundaries of a park in Central Boston.
}
\keyword{datasets}
