% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boston_sidewalk}
\alias{boston_sidewalk}
\title{Polygonal layer of sidewalks in Boston}
\format{A \code{SpatialLinesDataFrame} with 78 features.}
\usage{
boston_sidewalk
}
\description{
A \code{SpatialLinesDataFrame} object representing sidewalks in Central Boston.
}
\keyword{datasets}
