% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{build}
\alias{build}
\title{Polygonal layer of four buildings.}
\format{A \code{SpatialPolygonsDataFrame} with 4 features and 2 attributes:
\describe{
  \item{build_id}{Building ID}
  \item{BLDG_HT}{Building height, in meters}
}}
\usage{
build
}
\description{
A \code{SpatialPolygonsDataFrame} representing the outlines of four buildings. The attribute \code{BLDG_HT} contains building height, in meters.
}
\details{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}

