% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg2rad.R
\name{deg2rad}
\alias{deg2rad}
\title{Degrees to radians}
\usage{
deg2rad(deg)
}
\arguments{
\item{deg}{Angle in degrees}
}
\value{
\code{numeric} Angle in radians
}
\description{
Degrees to radians
}
\examples{
deg2rad(360) == 2*pi
}

