% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{elevation_as_mesh}
\alias{elevation_as_mesh}
\title{Convert Elevation Map as Triangle Polygon Mesh}
\usage{
elevation_as_mesh(elevation, texture, rotation, d)
}
\description{
Convert Elevation Map as Triangle Polygon Mesh
}
