% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{mesh_tile}
\alias{mesh_tile}
\alias{mesh_tri}
\title{Generate Meshes From Elevation Map}
\usage{
mesh_tile(L, dim)

mesh_tri(L, dim, order = FALSE)
}
\arguments{
\item{L}{list of equal length numeric vectors containing named elements 'x',
'y', and 'z'.  No NA elements allowed.}

\item{dim}{integer(2L) the dimensions of the elevation matrix that \code{L}
corresponds to.}
}
\value{
a list matrix of equal length numeric vectors, with rows each
representing a mesh element vertex, and as many columns as \code{L} has
elements.
}
\description{
\code{mesh_tile} produces a tile mesh, \code{mesh_tri} produces a triangle mesh.
The elevation map should have been converted to a list of equal length
vectors in "long" format.
}
