% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{persp_rel}
\alias{persp_rel}
\alias{persp_abs}
\title{Apply Relative Perspective Transformation}
\usage{
persp_rel(L, d)

persp_abs(L, d, fov)
}
\arguments{
\item{L}{list of equal length numeric vectors containing named elements 'x',
'y', and 'z'.  No NA elements allowed.}

\item{d}{scalar numeric observer distance from object along Z axis as a
multiple of the object's Z-depth, or for \code{persp_abs} the distance from the
midpoint of the model..}

\item{mode}{character in "rel", "abs". If "rel" then \code{d} is interpreted as
multiples of Z depth.  If "abs" it is taken directly as the distance from
the zero Z value, which should be the point around which the model was
rotated.}
}
\value{
L, with 'x', and 'y' components scaled for perspective, and
'x', 'y', and 'z' components normalized so that the origin is in
the middle of the 'x'-'y' with the observer Z position at zero.
}
\description{
Observer position is specified as a multiple of the Z depth of the object,
measured from the nearest Z point of the object.
}
