% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rad2deg.R
\name{rad2deg}
\alias{rad2deg}
\title{Radians to degrees}
\usage{
rad2deg(rad)
}
\arguments{
\item{rad}{Angle in radians}
}
\value{
\code{numeric} Angle in degrees
}
\description{
Radians to degrees
}
\examples{
rad2deg(2*pi) == 360
}

