% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slow-shade.R
\name{ray_shade1}
\alias{ray_shade1}
\title{Original Pure Base R Version of rayshader}
\usage{
ray_shade1(heightmap, anglebreaks = seq(40, 50, 1), sunangle = 315,
  maxsearch = 100)
}
\value{
numeric matrix of shadow intensities between with values in
&91;0,1&93;.
}
\description{
This is based on the original Tyler Morgan Wall \href{http://www.tylermw.com/throwing-shade/}{blog post}, with some clean-up to
functionalize, and with Wolf Vollprecht's bilinear interpolation
\href{https://nextjournal.com/wolfv/how-fast-is-r-with-fastr-pythran}{adaptation}
function.
}
\details{
It does not seem to work properly for all azimuths.  We have not bothered to
debug why. It seems to work for \link{0, 90}.

inheritParams rayshader::ray_shade
}
