% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pure-shade.R
\name{ray_shade2}
\alias{ray_shade2}
\alias{ray_shade3}
\title{Vectorized Base R Versions of \code{rayshader::ray_shade}}
\usage{
ray_shade2(heightmap, anglebreaks = seq(40, 50, 1), sunangle = 315,
  maxsearch = 100)

ray_shade3(heightmap, anglebreaks = seq(40, 50, 1), sunangle = 315,
  maxsearch = 100)
}
\arguments{
\item{heightmap}{A two-dimensional matrix, where each entry in the matrix is the elevation at that point. All points are assumed to be evenly spaced.}

\item{anglebreaks}{Default `seq(40,50,1)`. The azimuth angle(s), in degrees, as measured from the horizon from which the light originates.}

\item{sunangle}{Default `315` (NW). The angle, in degrees, around the matrix from which the light originates. Zero degrees is North, increasing clockwise.}

\item{maxsearch}{Default `100`. The maximum distance that the system should propogate rays to check. For longer}
}
\value{
numeric matrix of shadow intensities between with values in
&91;0,1&93;.
}
\description{
This is derived from the Wolf Vollprecht's
\href{https://nextjournal.com/wolfv/how-fast-is-r-with-fastr-pythran}{adaptation}
of the original pure R implementation from Tyler Morgan Wall's \href{http://www.tylermw.com/throwing-shade/}{blog post}.
}
\details{
\itemize{
\item ray_shade2: vectorized version with bilinear height interpolation
\item ray_shade3: vectorized version with closest point height approx
}
}
