% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render_elevation}
\alias{render_elevation}
\title{Render a 3D Elevation as a 2D Image}
\usage{
render_elevation(elevation, texture, rotation, resolution, d = 1,
  zord = "mesh", empty = 0)
}
\arguments{
\item{elevation}{numeric matrix of elevations; each matrix element is assumed
to be equally spaced in both X and Y dimensions.}

\item{texture}{a numeric matrix with values in \link{0,1} of the same dimensions
as \code{elevation}, where 1 means light and 0 means dark.  Currently only
grayscale textures are supported.}

\item{resolution}{integer(1L) the width of the output in pixels assuming
there is no rotation.  Actual width will depend on whether rotation causes
the width to change.}

\item{d}{numeric(1L) distance of the observer from nearest part of the model
as a multiple of the total depth of the model along the observation axis.}

\item{3}{x 3 numeric rotation matrix.}
}
\description{
\code{mrender_elevation} runs multiple projections with the same perspective and
field of view parameters so that they can be used together in a video.
}
