% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rishon}
\alias{rishon}
\title{Polygonal layer of four buildings in Rishon}
\format{A \code{SpatialPolygonsDataFrame} with 4 features and 2 attributes:
\describe{
  \item{build_id}{Building ID}
  \item{BLDG_HT}{Building height, in meters}
}}
\usage{
rishon
}
\description{
A \code{SpatialPolygonsDataFrame} object representing the outlines of four buildings located in Rishon-Le-Zion. The attribute \code{BLDG_HT} contains building height, in meters.
}
\keyword{datasets}

