% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-utils.R
\name{sequence2}
\alias{sequence2}
\title{Internally Vectorized Version of sequence}
\usage{
sequence2(nvec)
}
\arguments{
\item{nvec}{a non-negative integer vector each element of which
    specifies the end point of a sequence.}
}
\description{
Should be substantially faster for large \code{nvec}.
}
