% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadeFootprint.R
\name{shadeFootprint}
\alias{shadeFootprint}
\title{Shade footprint on the ground}
\usage{
shadeFootprint(build, height_field, sun_az, sun_elev, b = 0.01)
}
\arguments{
\item{build}{A \code{SpatialPolygonsDataFrame} object specifying the buildings outline.}

\item{height_field}{The name of the column with building height in \code{build}}

\item{sun_az}{Sun azimuth, in decimal degrees.}

\item{sun_elev}{Sun elevation, in decimal degrees.}

\item{b}{Buffer size for shade footprints of individual segments of a given polygon; used to eliminate minor internal holes in the resulting shade polygon.}
}
\description{
Creates a vector layer of shade footprint on the ground, given sun position and extruded obstacles (usually a buildings layer).
The calculation method was inspired by the of Morel Weisthal's MSc thesis at Ben-Gurion University.
}
\examples{
data(build)
footprint = shadeFootprint(build, "BLDG_HT", 310, 30)
plot(footprint, col = adjustcolor("lightgrey", alpha.f = 0.5))
plot(build, add = TRUE, col = "darkgrey")

}
\references{
Weisthal, M. (2014). Assessment of potential energy savings in Israel through climate-aware residential building design (Doctoral dissertation, Ben-Gurion University of the Negev).
\url{http://aranne5.bgu.ac.il/others/WeisthalMorel.pdf}
}

