% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadeImageWall.R
\name{shadeImageWall}
\alias{shadeImageWall}
\title{Title}
\usage{
shadeImageWall(seg, seg_height_field, build, build_height_field, solar_pos,
  sample_dist = 1, shift_dist = 0.01, messages = TRUE)
}
\arguments{
\item{messages}{}
}
\description{
Title
}

