% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadePropWall.R
\name{shadePropWall}
\alias{shadePropWall}
\title{Shaded proportion of building walls}
\usage{
shadePropWall(seg, seg_height_field, build, build_height_field, sun_az,
  sun_elev, sample_dist = 1, shift_dist = 0.01)
}
\arguments{
\item{seg}{A \code{SpatialLinesDataFrame} object where each feature is a single segment representing a wall.}

\item{seg_height_field}{The name of the column with wall height in \code{seg}}

\item{build}{A \code{SpatialPolygonsDataFrame} object specifying the buildings outline}

\item{sun_az}{Sun azimuth, in decimal degrees.}

\item{sun_elev}{Sun elevation, in decimal degrees.}

\item{shift_dist}{The distance for shifting the examined locations away from wall to avoid self-shading. Default is 1 cm.}

\item{height_field}{The name of the column with building height in \code{build}}
}
\description{
Shaded proportion of building walls
}
\examples{
data(build)
seg = toSeg(build[1, ])
plot(seg, col = sample(rainbow(length(seg))))
shadePropWall (
  seg = seg,
  seg_height_field = "BLDG_HT",
  build = build,
  build_height_field = "BLDG_HT",
  sun_az = 30,
  sun_elev = 20,
  sample_dist = 1,
  shift_dist = 0.01
)
raster::text(rgeos::gCentroid(seg, byid = TRUE), 1:length(seg))

}

