% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunElevAz.R
\name{sunElevAz}
\alias{sunElevAz}
\title{Find sun azimuth and elevation angles}
\usage{
sunElevAz(timeGMT, lon, lat)
}
\arguments{
\item{timeGMT}{Time, a \code{POSIXct} object.}

\item{lon}{Longitude.}

\item{lat}{Latitude.}
}
\value{
A list with two components, \code{elevation} and \code{azimuth}, both in decimal degrees.
}
\description{
The function returns sun azimuth and elevation given time and spatial location.
}
\details{
This function is a modified version of the following 'stackoverflow' forum post by
Josh O'Brien:
\url{http://stackoverflow.com/questions/8708048/position-of-the-sun-given-time-of-day-latitude-and-longitude}
}
\examples{

time_local = as.POSIXct("1999-01-01 12:00:00", tz = "Asia/Jerusalem")
time_GMT = as.POSIXct(format.POSIXct(time_local, tz = "GMT"), tz = "GMT")
sunElevAz(time_GMT, lat = 31.974447, lon = 34.791708)

}
\references{
Michalsky, J.J. 1988. The Astronomical Almanac's algorithm for approximate solar position (1950-2050). Solar Energy. 40(3):227-235.

Michalsky, J.J. 1989. Errata. Solar Energy. 43(5):323.

Spencer, J.W. 1989. Comments on "The Astronomical Almanac's Algorithm for Approximate Solar Position (1950-2050)". Solar Energy. 42(4):353.

Walraven, R. 1978. Calculating the position of the sun. Solar Energy. 20:393-397.
}

