% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunLocation.R
\name{sunLocation}
\alias{sunLocation}
\title{Find sun location relatively to the given point}
\usage{
sunLocation(location, sun_az, sun_elev)
}
\arguments{
\item{location}{A \code{SpatialPoints*} object specifying viewer location.}

\item{sun_az}{Sun azimuth, in decimal degrees.}

\item{sun_elev}{Sun elevation, in decimal degrees.}
}
\value{
A \code{SpatialPointsDataFrame} object specifying sun location.
  The \code{hight} field specifies sun *height* above projection plane.
}
\description{
The function returns a point layer with the sun position projected on the CRS plain.
}
\examples{
 sunLocation(rgeos::gCentroid(build), 45, 60)
}

