% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunPosition.R
\name{sunPosition}
\alias{sunPosition}
\title{Find sun azimuth and elevation angles}
\usage{
sunPosition(timeGMT, lon, lat)
}
\arguments{
\item{timeGMT}{Time, a \code{POSIXct} object.}

\item{lon}{Longitude.}

\item{lat}{Latitude.}
}
\value{
A list with two components, \code{elevation} and \code{azimuth}, both in decimal degrees.
}
\description{
The function returns sun azimuth and elevation given time and spatial location.
}
\examples{
 time_local = as.POSIXct("1999-01-01 12:00:00", tz = "Asia/Jerusalem")
 time_GMT = as.POSIXct(format.POSIXct(time_local, tz = "GMT"), tz = "GMT")
 sunPosition(time_GMT, lat = 31.974447, lon = 34.791708)
}

