% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toGMT.R
\name{toGMT}
\alias{toGMT}
\title{Local time to GMT}
\usage{
toGMT(time)
}
\arguments{
\item{time}{Time, a \code{POSIXct} object.}
}
\value{
A a \code{POSIXct} object, in GMT.
}
\description{
The function transforms a \code{POSIXct} object in any given time zone to GMT.
}
\examples{
time = as.POSIXct("1999-01-01 12:00:00", tz = "Asia/Jerusalem")
toGMT(time)

}

