% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVF.R, R/toSeg.R
\name{toSeg}
\alias{toSeg}
\title{Sky View Factor (SVF) calculation}
\usage{
toSeg(x)

toSeg(x)
}
\arguments{
\item{x}{A \code{SpatialLines*} or a \code{SpatialPolygons*} object.}

\item{x}{A \code{SpatialLines*} or a \code{SpatialPolygons*} object.}
}
\value{
A \code{SpatialLines} object where each segment is represented by a separate feature.

A \code{SpatialLines} object where each segment is represented by a separate feature.
}
\description{
Split lines or polygons to separate segments.

Split lines or polygons to separate segments.
}
\examples{
data(build)
seg = toSeg(build[1, ])
plot(seg, col = sample(rainbow(length(seg))))
raster::text(rgeos::gCentroid(seg, byid = TRUE), 1:length(seg))

data(build)
seg = toSeg(build[1, ])
plot(seg, col = sample(rainbow(length(seg))))
raster::text(rgeos::gCentroid(seg, byid = TRUE), 1:length(seg))

}
\references{
This function uses a modified version of code from the following 'r-sig-geo' post by Roger Bivand:
\url{https://stat.ethz.ch/pipermail/r-sig-geo/2013-April/017998.html}

This function uses a modified version of code from the following 'r-sig-geo' post by Roger Bivand:
\url{https://stat.ethz.ch/pipermail/r-sig-geo/2013-April/017998.html}
}

