% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{beersheva_elev}
\alias{beersheva_elev}
\title{DEM of Ramot neighborhood, Beer-Sheva}
\format{
A \code{RasterLayer} representing a grid of 1974 raster cells, each cell is a 30*30 meters rectangle. Data source is the Shuttle Radar Topography Mission (SRTM) 1 Arc-Second Global dataset.
}
\usage{
beersheva_elev
}
\description{
Digital Elevation Model (DEM) of Ramot neighborhood, Beer-Sheva. Raster values represent elevation above sea level, in meters.
}
\references{
https://lta.cr.usgs.gov/SRTM1Arc
}
\keyword{datasets}
