% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadeImageWall.R
\name{shadeImageWall}
\alias{shadeImageWall}
\title{Title}
\usage{
shadeImageWall(seg, seg_height_field, build, build_height_field, solar_pos,
  sample_dist = 1, shift_dist = 0.01, messages = TRUE)
}
\arguments{
\item{messages}{}
}
\description{
Title
}
\examples{

# data(build)
time = as.POSIXct("2004-12-24 11:30:00", tz = "Asia/Jerusalem")
solar_pos = maptools::solarpos(
  matrix(c(34.7767978098526, 31.9665936050395), ncol = 2),
  time
)
seg = shadow::toSeg(build[2, ])[5, ]

plot(build)
plot(seg, add = TRUE, col = "red")

seg = seg
  seg_height_field = "BLDG_HT"
  build = build
  build_height_field = "BLDG_HT"
  solar_pos = solar_pos
  sample_dist = 1
  shift_dist = 0.01

shadeImageWall(
  seg,
  seg_height_field = "BLDG_HT",
  build = build,
  build_height_field = "BLDG_HT",
  solar_pos = solar_pos,
  sample_dist = 1,
  shift_dist = 0.01
)
}

