% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadePropWall.R
\name{shadePropWall}
\alias{shadePropWall}
\title{Shaded proportion of building walls}
\usage{
shadePropWall(seg, seg_height_field, build, build_height_field, sun_az,
  sun_elev, sample_dist = 1, shift_dist = 0.01)
}
\arguments{
\item{seg}{A \code{SpatialLinesDataFrame} object where each feature is a single segment representing a wall.}

\item{seg_height_field}{The name of the column with wall height in \code{seg}}

\item{build}{A \code{SpatialPolygonsDataFrame} object specifying the buildings outline}

\item{sun_az}{Sun azimuth, in decimal degrees.}

\item{sun_elev}{Sun elevation, in decimal degrees.}

\item{shift_dist}{The distance for shifting the examined locations away from wall to avoid self-shading. Default is 1 cm.}

\item{height_field}{The name of the column with building height in \code{build}}
}
\value{
Proportion of shaded area of the given segment.
}
\description{
Calculates the shaded proportion of a segment (usually representing a building wall),
given the segment vertical height as well as the outlines and heights of obstacles
(usually buildings).
}
\details{
#' @note For a correct geometric calculation, make sure that:\itemize{
\item{The layers \code{location} and \code{build} are projected}
\item{The values in \code{height_field} of \code{build} are given in the same distance units as the CRS (e.g. meters when using UTM)}
}
}
\examples{
data(build)
seg = shadow::toSeg(build[2, ])

# Calculate shade proportion for walls of 2nd feature from 'build'
props = shadePropWall(
  seg = seg,
  seg_height_field = "BLDG_HT",
  build = build,
  build_height_field = "BLDG_HT",
  sun_az = 330,
  sun_elev = 10,
  sample_dist = 1,
  shift_dist = 0.01
)

# Plot
cols = grey(seq(1, 0.5, length.out = 100))
props = round(props, 2) * 100
plot(build)
plot(seg, add = TRUE, col = cols[props], lwd = 2)
raster::text(rgeos::gCentroid(seg, byid = TRUE), props, cex = 0.75)

}

